requirements.txt
# Core Dependencies
streamlit==1.32.0
python-docx==1.1.2
requests==2.31.0
python-box==7.1.1
python-dateutil==2.9.0.post0
pytz==2024.1

# Box SDK and Authentication
boxsdk==3.14.0  # Latest version compatible with Python 3.12
python-jose==3.3.0
cryptography==42.0.5
PyJWT==2.8.0

# AI and ML (if using AI features)
openai==1.12.0
tiktoken==0.5.2

# Data Processing
pandas==2.2.1
numpy==1.26.4

# Development and Testing
pytest==8.1.1
pytest-cov==4.1.0
black==24.3.0
isort==5.13.2
flake8==7.0.0
mypy==1.8.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.3

# Utilities
python-dotenv==1.0.1
tqdm==4.66.1
loguru==0.7.2
